+++ src/main/kotlin/com/zepben/evolve/services/network/tracing/feeder/SetDirection.kt
@@ -11,6 +11,7 @@ package com.zepben.evolve.services.network.tracing.feeder
 import com.zepben.evolve.cim.iec61970.base.core.Feeder
 import com.zepben.evolve.cim.iec61970.base.core.Terminal
 import com.zepben.evolve.cim.iec61970.base.wires.BusbarSection
+import com.zepben.evolve.cim.iec61970.base.wires.Cut
 import com.zepben.evolve.cim.iec61970.base.wires.PowerTransformer
 import com.zepben.evolve.services.network.NetworkService
 import com.zepben.evolve.services.network.tracing.networktrace.NetworkTrace
@@ -36,11 +37,12 @@ class SetDirection {
             return FeederDirection.CONNECTOR
         }
 
-        val directionApplied = step.data
-        val nextDirection = when (directionApplied) {
-            FeederDirection.UPSTREAM -> FeederDirection.DOWNSTREAM
-            FeederDirection.DOWNSTREAM, FeederDirection.CONNECTOR -> FeederDirection.UPSTREAM
-            else -> FeederDirection.NONE
+        val nextDirection = when {
+            step.data == FeederDirection.NONE -> FeederDirection.NONE
+            nextPath.tracedInternally -> FeederDirection.DOWNSTREAM
+            nextPath.toEquipment is Cut -> FeederDirection.UPSTREAM
+            nextPath.didTraverseAcLineSegment -> FeederDirection.DOWNSTREAM
+            else -> FeederDirection.UPSTREAM
         }
 
         //