+++ src/main/kotlin/com/zepben/evolve/services/network/tracing/networktrace/NetworkTraceQueueNext.kt
@@ -9,21 +9,22 @@
 package com.zepben.evolve.services.network.tracing.networktrace
 
 import com.zepben.evolve.annotations.ZepbenExperimental
+import com.zepben.evolve.services.network.tracing.networktrace.operators.NetworkStateOperators
 import com.zepben.evolve.services.network.tracing.traversal.StepContext
 import com.zepben.evolve.services.network.tracing.traversal.Traversal
 
-internal abstract class NetworkTraceQueueNext(private val pathProvider: NetworkTraceStepPathProvider) {
+internal abstract class NetworkTraceQueueNext(private val stateOperators: NetworkStateOperators) {
 
     internal class Basic<T> : NetworkTraceQueueNext, Traversal.QueueNext<NetworkTraceStep<T>> {
 
         private val getNextSteps: (item: NetworkTraceStep<T>, context: StepContext) -> Iterator<NetworkTraceStep<T>>
 
-        constructor(pathProvider: NetworkTraceStepPathProvider, computeData: ComputeData<T>) : super(pathProvider) {
+        constructor(stateOperators: NetworkStateOperators, computeData: ComputeData<T>) : super(stateOperators) {
             getNextSteps = { item, context -> nextTraceSteps(item, context, computeData).iterator() }
         }
 
         @ZepbenExperimental
-        constructor(pathProvider: NetworkTraceStepPathProvider, computeData: ComputeDataWithPaths<T>) : super(pathProvider) {
+        constructor(stateOperators: NetworkStateOperators, computeData: ComputeDataWithPaths<T>) : super(stateOperators) {
             getNextSteps = { item, context -> nextTraceSteps(item, context, computeData).iterator() }
         }
 
@@ -36,12 +37,12 @@ internal abstract class NetworkTraceQueueNext(private val pathProvider: NetworkT
 
         private val getNextSteps: (item: NetworkTraceStep<T>, context: StepContext) -> List<NetworkTraceStep<T>>
 
-        constructor(pathProvider: NetworkTraceStepPathProvider, computeData: ComputeData<T>) : super(pathProvider) {
+        constructor(stateOperators: NetworkStateOperators, computeData: ComputeData<T>) : super(stateOperators) {
             getNextSteps = { item, context -> nextTraceSteps(item, context, computeData).toList() }
         }
 
         @ZepbenExperimental
-        constructor(pathProvider: NetworkTraceStepPathProvider, computeData: ComputeDataWithPaths<T>) : super(pathProvider) {
+        constructor(stateOperators: NetworkStateOperators, computeData: ComputeDataWithPaths<T>) : super(stateOperators) {
             getNextSteps = { item, context -> nextTraceSteps(item, context, computeData) }
         }
 
@@ -62,7 +63,7 @@ internal abstract class NetworkTraceQueueNext(private val pathProvider: NetworkT
         computeData: ComputeData<T>,
     ): Sequence<NetworkTraceStep<T>> {
         val nextNumTerminalSteps = currentStep.nextNumTerminalSteps()
-        return pathProvider.nextPaths(currentStep.path).map {
+        return stateOperators.nextPaths(currentStep.path).map {
             val data = computeData.computeNext(currentStep, currentContext, it)
             NetworkTraceStep(it, nextNumTerminalSteps, it.nextNumEquipmentSteps(currentStep.numEquipmentSteps), data)
         }
@@ -75,7 +76,7 @@ internal abstract class NetworkTraceQueueNext(private val pathProvider: NetworkT
         computeNextT: ComputeDataWithPaths<T>,
     ): List<NetworkTraceStep<T>> {
         val nextNumTerminalSteps = currentStep.nextNumTerminalSteps()
-        val nextPaths = pathProvider.nextPaths(currentStep.path).toList()
+        val nextPaths = stateOperators.nextPaths(currentStep.path).toList()
         return nextPaths.map {
             val data = computeNextT.computeNext(currentStep, currentContext, it, nextPaths)
             NetworkTraceStep(it, nextNumTerminalSteps, it.nextNumEquipmentSteps(currentStep.numEquipmentSteps), data)