+++ src/main/kotlin/com/zepben/evolve/services/network/tracing/networktrace/operators/NetworkStateOperators.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Zeppelin Bend Pty Ltd
+ * Copyright 2025 Zeppelin Bend Pty Ltd
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -9,6 +9,8 @@
 package com.zepben.evolve.services.network.tracing.networktrace.operators
 
 import com.zepben.evolve.services.network.tracing.networktrace.NetworkTrace
+import com.zepben.evolve.services.network.tracing.networktrace.NetworkTraceStep
+import com.zepben.evolve.services.network.tracing.networktrace.NetworkTraceStepPathProvider
 
 /**
  * Interface providing access to and operations on specific network state properties and functions for items within a network.
@@ -28,7 +30,8 @@ interface NetworkStateOperators :
     FeederDirectionStateOperations,
     EquipmentContainerStateOperators,
     InServiceStateOperators,
-    PhaseStateOperators {
+    PhaseStateOperators,
+    ConnectivityStateOperators {
 
     companion object {
         /**
@@ -50,11 +53,27 @@ private class NormalNetworkStateOperators : NetworkStateOperators,
     FeederDirectionStateOperations by FeederDirectionStateOperations.NORMAL,
     EquipmentContainerStateOperators by EquipmentContainerStateOperators.NORMAL,
     InServiceStateOperators by InServiceStateOperators.NORMAL,
-    PhaseStateOperators by PhaseStateOperators.NORMAL
+    PhaseStateOperators by PhaseStateOperators.NORMAL {
+
+    private val networkTraceStepPathProvider = NetworkTraceStepPathProvider(this)
+
+    // This is not implemented in the same NORMAL/CURRENT single instance pattern as all the other
+    // StateOperator interfaces because the implementation of next paths has a dependency on other state operators.
+    override fun nextPaths(path: NetworkTraceStep.Path): Sequence<NetworkTraceStep.Path> =
+        networkTraceStepPathProvider.nextPaths(path)
+}
 
 private class CurrentNetworkStateOperators : NetworkStateOperators,
     OpenStateOperators by OpenStateOperators.CURRENT,
     FeederDirectionStateOperations by FeederDirectionStateOperations.CURRENT,
     EquipmentContainerStateOperators by EquipmentContainerStateOperators.CURRENT,
     InServiceStateOperators by InServiceStateOperators.CURRENT,
-    PhaseStateOperators by PhaseStateOperators.CURRENT {}
+    PhaseStateOperators by PhaseStateOperators.CURRENT {
+
+    private val networkTraceStepPathProvider = NetworkTraceStepPathProvider(this)
+
+    // This is not implemented in the same NORMAL/CURRENT single instance pattern as all the other
+    // StateOperator interfaces because the implementation of next paths has a dependency on other state operators.
+    override fun nextPaths(path: NetworkTraceStep.Path): Sequence<NetworkTraceStep.Path> =
+        networkTraceStepPathProvider.nextPaths(path)
+}